finished

+++ tianya.py
@@ -3,11 +3,21 @@
 import urllib,urllib2, re, lxml, urlparse, os, md5, time
 from lxml import etree
 from StringIO import StringIO
-
+from HTMLParser import HTMLParser
+import thread, threading, sys
 
 # 地缘看世界URL
 base_url = 'http://bbs.tianya.cn/post-worldlook-223829-1.shtml'
 url_template = 'http://bbs.tianya.cn/post-worldlook-223829-%d.shtml'
+
+html_parser = HTMLParser()
+pages = []
+pages_lock = thread.allocate_lock()
+
+
+def html_parse(html):
+	return html_parser.unescape(html)
+	
 def get_page_url(page):
 	return (url_template % page)
 
@@ -24,7 +34,7 @@ def get_data_from_url(url):
 		data = urllib2.urlopen(req).read()
 	except Exception:
 		data = ''
-	return 
+	return data
 
 def get_dom_from_html(html):
 	return etree.parse(StringIO(html), etree.HTMLParser())
@@ -35,7 +45,7 @@ def get_dom_from_url(url):
 	return T
 	
 def dom_to_html(T):
-	return etree.tostring(T, pretty_print=True, method="html")
+	return etree.tostring(T, pretty_print=True)
 
 def parse_post_div(post_div):
 	# print post_div.attrib
@@ -45,22 +55,25 @@ def parse_post_div(post_div):
 	else:
 		author = '鄙视抢沙发的'.decode('utf-8')
 		datetime = '2009-07-12 12:00:00'
+	author = html_parse(author)
 	content = post_div.find('.//div[@class="bbs-content"]')
 	if content is None:
 		content = post_div.find('.//div[@class="bbs-content clearfix"]')
 	# print content
 	content = dom_to_html(content)
 	
-	content = preprocess_post(author, datetime, content)
+	content = html_parse(preprocess_post(author, datetime, content))
 	return author, datetime, content
 def save_file_from_url(url, path):
+	if os.path.exists(path):
+		return
 	data = get_data_from_url(url)
 	f = open(path, 'wb')
 	f.write(data)
 	f.close()
 	
 def get_local_single_post_path(author, datetime):
-	fname = datetime[:10] +'-'+ author 
+	fname = re.sub(r'[\s\:]','-',datetime) +'-'+ author 
 	fname = re.sub(r'[:\\\/\?\*\"\<\>\|\r\n]',' ', fname)
 	path = './data/%s.html' % fname.encode('gbk','ignore')
 	return path
@@ -70,11 +83,13 @@ def preprocess_post(author, datetime, content):
 	imgs = T.findall('.//img')
 	if imgs is not None:
 		for img in imgs:
-			img_path = './data/images/'+md5.new(img.attrib['original']).hexdigest()+'.jpg'
-			save_file_from_url(img.attrib['original'], './images/'+md5.new(img.attrib['original']).hexdigest()+'.jpg')
-			print '[saved]', img.attrib['original']
+			img_name = md5.new(img.attrib['original'].encode('utf-8','ignore')).hexdigest()+'.jpg'
+			img_path = './data/images/'+img_name
 			
-			content = content.replace(img.attrib['src'], img_path)
+			save_file_from_url(img.attrib['original'], img_path)
+			print '[saved]', img.attrib['original'], img_path
+			
+			content = content.replace(img.attrib['src'], './images/'+img_name)
 			
 	content = '<p>%s %s</p><div>%s</div>' % (author, datetime, content)
 	return content
@@ -96,27 +111,55 @@ def save_page(page):
 	fpage = open(page_path, 'wb')
 	for div in post_divs:
 		author, datetime, content = parse_post_div(div)
-		print author, datetime
-		if author is '鄙视抢沙发的'.decode('utf-8'):
+		# print author, datetime
+		# print repr('鄙视抢沙发的'.decode('utf-8'))
+		# print repr(html_parse(author))
+		if author == '鄙视抢沙发的'.decode('utf-8'):
 			post_path = get_local_single_post_path(author, datetime)
-			write_to_file(content, post_path)
-			print '[saved]', post_path.encode('gbk','ignore')
-		fpage.write(content.encode('utf-8') + '\r\n')
+			if os.path.exists(post_path):
+				continue
+			write_to_file(content.encode('utf-8','ignore'), post_path)
+			print '[saved]', post_path
+		fpage.write(content.encode('utf-8','ignore') + '\r\n')
 	
 	fpage.close()
 	print '[saved]', page_path
 	time.sleep(5)
 	
+def get_visisted_urls():
+	return open('urls.txt','r').read().split('\n')
+def save_visited_url(url):
+	f = open('urls.txt','a')
+	f.write(url + '\n')
+	f.close()
+
+def crawl():
+	pages_lock.acquire()
+	if len(pages)==0:
+		pages_lock.release()
+		print '[exit]', 'threading exit!'
+		return
+	i = pages.pop(0)
+	pages_lock.release()
 	
-def main():
+	url = get_page_url(i)
+	if url not in visited_urls:		
+		save_page(i)
+		save_visited_url(url)
+	crawl()
+	
+
+	
+if __name__ == '__main__':
 	T = get_dom_from_url(base_url)
 	page_num = T.find('//div[@class="atl-pages"]/form/a[last()-1]').text
 	page_num = int(page_num)
-	for i in range(1,page_num+1):
-		save_page(i)
+	visited_urls = get_visisted_urls()
+	pages = range(1,page_num+1)
 	
 	
-if __name__ == '__main__':
-	main()
+	for i in range(10):
+		threading.Thread(target=crawl).start()
+
